ログインしているユーザーと異なるコメントの削除と変更をとめるif文追加

+++ boardservice/delete.php
@@ -4,23 +4,39 @@ if ($_POST['action'] == 'cancel') {
 exit;
 }
 require_once 'db_connect.php';
-try{
-    $dbh = db_connect();
+require_once 'function.php';
 
-    // 該当するIDからコメント削除
-    $id = $_POST['id'];
-    $sql = 'DELETE FROM comments WHERE id = :id';
-    $pre = $dbh->prepare($sql);
+session_start();
+unlogined_session();
 
-    $pre->bindValue(':id', $id);
+// var_dump($_POST);
+// echo "<br>";
+// var_dump($_SESSION);
 
-    $r = $pre->execute();
-} catch (PDOException $e) {
-    echo "エラーが発生。最初から再度IDを選択。 (" , $e->getMessage() , ")";
-    return ;
 
+// ログインされていれば　$_SESSION['id']とusers.id　が同じになる
+// delete_regi.phpから異なるユーザーがアクセスしたときにcomments.user_id と users.idが一致するか判別
+if($_SESSION['id'] == $_POST['user_id']){
+    try{
+        $dbh = db_connect();
+
+        // 該当するIDからコメント削除
+        $id = $_POST['id'];
+        $sql = 'DELETE FROM comments WHERE id = :id';
+        $pre = $dbh->prepare($sql);
+
+        $pre->bindValue(':id', $id);
+
+        $r = $pre->execute();
+      } catch (PDOException $e) {
+        echo "エラーが発生。最初から再度IDを選択。 (" , $e->getMessage() , ")";
+        return ;
+
+    }
+} else {
+  header('Location: index.php');
+  exit;
 }
-
 ?>
 <!DOCTYPE html>
 <html lang="ja">
